switch to double precision

+++ src/buildings/tests/buildings_test_case.rs
@@ -3,7 +3,7 @@ use crate::sectors::SectorsInputs;
 use crate::sectors::SectorsRawValues;
 
 impl SectorsInputs{
-    pub fn set(&mut self, private: f32, industry:f32, schools: f32, public: f32){
+    pub fn set(&mut self, private: f64, industry:f64, schools: f64, public: f64){
         self.private.set_values(private);
         self.industry.set_values(industry);
         self.schools.set_values(schools);
