+++ src/result.rs
@@ -1,7 +1,7 @@
 
 pub struct Results{
     pub id: String,
-    values: Vec<f32>,
+    values: Vec<f64>,
     start_year: u32,
 }
 
@@ -15,24 +15,24 @@ impl Results{
         }
     }
 
-    pub fn get_year(&self, year: u32) -> f32{
+    pub fn get_year(&self, year: u32) -> f64{
         return self.values[(year-self.start_year) as usize];
     }
 
-    pub fn sum_years(&self, start_year: u32, end_year: u32) -> f32{
+    pub fn sum_years(&self, start_year: u32, end_year: u32) -> f64{
         return self.values
             [(start_year-self.start_year) as usize
             .. (end_year-self.start_year+1) as usize]
             .iter().sum();
     }
 
-    pub fn set_year_value(&mut self, year: u32, value: f32){
+    pub fn set_year_value(&mut self, year: u32, value: f64){
         let year_index = (year - self.start_year) as usize;
         self.values[year_index] = value;
     }
 
 
-    pub fn get_values(& self) -> &Vec<f32>{
+    pub fn get_values(& self) -> &Vec<f64>{
         return & self.values;
     }
 }
